Enhanced version of finalfit::summary_factorlist

Generate aggregated summary statistics and tests for categorical and continuous variables, with an option for Fisher's correction in contingency tables.

This function extends the functionality of the `summary_factorlist` function by introducing Fisher's test and Shapiro-Wilk test. Normality distributions of continuous variables is evaluated by Shapiro-Wilk test and required test is decided based on the test. For categorical variables, expected cell frequencies is calculated (expected cell frequencies less than 5 in more than 20 contingency table may require using Fisher test) and required test (Chi-square or Fisher) is decided based on this. row_col_sums function can be used for presenting different sum styles. In the default one, the sums of the rows is 100 This approach is good for summary tables.

Usage

ag_ff_summary(
  dataset,
  strata = NULL,
  table_vars,
  row_col_sums = "row_col_based",
  all_cont = "shapiro",
  manual_nonparams = NULL,
  fisher_correction = TRUE,
  na_include = FALSE,
  summary_factorlist_args = list(),
  ...
)

Arguments

dataset

A data frame containing the variables of interest.

strata

The dependent variable (outcome) in the analysis, should be a factor.

table_vars

A character vector specifying the explanatory variables (predictors) for the analysis. These can be factors, characters, or numerics.

row_col_sums

A character string specifying the type of summary statistics to be generated. Options include "row_based" for row-based summaries, "col_based" for column-based summaries, and "row_col_based" for both row and column-based summaries. Default is "row_col_based".

all_cont

A character string specifying the approach for continuous variables. Options include "parametric" for parametric tests (mean), "nonparametric" for non-parametric tests (median), and "shapiro" for using the Shapiro-Wilk test for normality and choosing between parametric and non-parametric accordingly. Default is "shapiro". Manual option is also possible with manual_nonparams.

manual_nonparams

If all_cont is set to "manual", this argument should contain the values (not index, the name of the variable) for cont_nonpara. It cannot be empty in this case.

fisher_correction

Logical, indicating whether Fisher's correction should be applied for expected cell frequencies less than 5 in more than 20 Default is TRUE.

na_include

Logical, indicating whether missing values should be included in the analysis. Default is TRUE.

summary_factorlist_args

A list containing additional arguments to be passed to ag_ff_summary. All comes from the finalfit::summaryfactorlist function. Default is an empty list.

...

Additional arguments to be passed to ag_ff_summary.

Value

A data frame containing the aggregated summary statistics and test results.

See also

Author

Ali Guner

Examples

if (FALSE) {
# Example usage of ag_ff_summary function
library(palmerpenguins)

strata <- "species"
table_vars_1 <- penguins %>%
dplyr::select(-species) %>%
names()

ag_ff_summary(
dataset = penguins,
table_vars = table_vars_1,
strata = "species"
)
}